Adjustments to pedido

+++ JS/controllerProductosPedidos.js
@@ -7,44 +7,65 @@ crearListaProductosPedidos();
 
 function crearListaProductosPedidos(){
 
+
     var pedidoRecuperado = localStorage.getItem("pedidoDelComensal");
     productosPedidos = JSON.parse(pedidoRecuperado);
 
-    // var productosRecuperados = localStorage.getItem("informacionProductos");
-    // productosRestaurante = JSON.parse(productosRecuperados);
+    if (productosPedidos.length < 1) {
 
-    var divProductosPedidos = document.getElementById("productosPedidos");
+        alert ("No hay ningún artículo en pedido");
+        location.href=`list.php?restaurante=${restaurante}&idComensal=${idComensal}`;
+    }
+    else {
 
-    for (let i = 0; i < productosPedidos.length; i++){
+        var divProductosPedidos = document.getElementById("productosPedidos");
 
-        var divProductoPedido = document.createElement("div");
-        divProductoPedido.className = "itemsPedidos";
-        divProductosPedidos.appendChild(divProductoPedido);
-
-        var nombreProductoPedido = document.createElement("p");
-        nombreProductoPedido.className = "productoPedidoNombre";
-        nombreProductoPedido.textContent = productosPedidos[i].nombre;
-        divProductoPedido.appendChild(nombreProductoPedido);
-        
-        var cantidadProductoPedido = document.createElement("p");
-        cantidadProductoPedido.className = "cantidadProductoPedido";
-        cantidadProductoPedido.textContent = productosPedidos[i].cantidadProducto;
-        divProductoPedido.appendChild(cantidadProductoPedido);
-
-        var precioProductoPedido = document.createElement("p");
-        precioProductoPedido.className = "precioProductoPedido";
-        precioProductoPedido.textContent = `$ `+ (parseFloat(productosPedidos[i].precio) * parseFloat(productosPedidos[i].cantidadProducto));
-        divProductoPedido.appendChild(precioProductoPedido);
-
-        var eliminarProducto = document.createElement("button");
-        eliminarProducto.textContent = "Eliminar";
-        eliminarProducto.id = "botonEliminar";
-        eliminarProducto.value = i;
-        eliminarProducto.onclick = function (){eliminarProductoDePedido(this.value);};
-        divProductoPedido.appendChild(eliminarProducto);
+        for (let i = 0; i < productosPedidos.length; i++){
+    
+            var divProductoPedido = document.createElement("div");
+            divProductoPedido.className = "itemsPedidos";
+            divProductosPedidos.appendChild(divProductoPedido);
+    
+            
+            var divInfoProducto = document.createElement("div");
+            divInfoProducto.className = "productoPedidoInfo";
+            divProductoPedido.appendChild(divInfoProducto);
+    
+            var nombreProductoPedido = document.createElement("p");
+            nombreProductoPedido.className = "productoPedidoNombre";
+            nombreProductoPedido.textContent = productosPedidos[i].nombre;
+            divInfoProducto.appendChild(nombreProductoPedido);
+            
+            var cantidadProductoPedido = document.createElement("p");
+            cantidadProductoPedido.className = "cantidadProductoPedido";
+            cantidadProductoPedido.textContent = `x`+ productosPedidos[i].cantidadProducto;
+            divInfoProducto.appendChild(cantidadProductoPedido);
+    
+            var precioProductoPedido = document.createElement("p");
+            precioProductoPedido.className = "precioProductoPedido";
+            precioProductoPedido.textContent = `$ `+ (parseFloat(productosPedidos[i].precio) * parseFloat(productosPedidos[i].cantidadProducto));
+            divInfoProducto.appendChild(precioProductoPedido);
+    
+            var divBotonProducto = document.createElement("div");
+            divBotonProducto.className = "botonProducto";
+            divProductoPedido.appendChild(divBotonProducto);
+    
+            var eliminarProducto = document.createElement("button");
+            eliminarProducto.textContent = "Eliminar";
+            eliminarProducto.id = "botonEliminar";
+            eliminarProducto.className = "botonEliminar";
+            eliminarProducto.value = i;
+            eliminarProducto.onclick = function (){eliminarProductoDePedido(this.value);};
+            divBotonProducto.appendChild(eliminarProducto);
+    
+        }
+    
 
     }
 
+   
+
+   
 }
 
 function eliminarProductoDePedido(idProducto){
